chore() :move models schema to snu-lib
