+++ packages/lib/package.json
@@ -35,6 +35,7 @@
   },
   "dependencies": {
     "date-fns-tz": "^3.1.3",
+    "mongoose": "^6.12.9",
     "query-string": "^6.14.1",
     "sanitize-html": "^2.12.1"
   },
