+++ packages/lib/src/mongoSchema/index.ts
@@ -1,0 +1,3 @@
+export type InterfaceExtended<T> = T & { _id: string; _doc?: T };
+
+export * from "./cle/classe";
